move resolver to its own file
